[CONTROLLER-DB]split db migration
**Phenomenon and reproduction steps**

**Root cause and solution**

- add check of master/slave region
- add error tolerance for sidecar check
- split for module replace

**Impactions**

**Test method**

**Affected branch(es)**

* main

**Checklist**

- [ ] Dependencies update required
- [ ] Common bug (similar problem in other repo)
